requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
b2sdk==2.5.1
Babel==2.14.0
boto3==1.28.12
botocore==1.31.12
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==4.2.7
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-phonenumber-field==7.3.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
logfury==1.0.1
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.28
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.20
whitenoise==6.6.0
